extra ones of merge

+++ include/cute/arch/xe_copy_1B.hpp
@@ -140,16 +140,16 @@ SYCL_DEVICE_BUILTIN(void __builtin_IB_subgroup_block_write_flat_u8_m8k16v2(
 
 // U8 prefetch
 SYCL_DEVICE_BUILTIN(void __builtin_IB_subgroup_block_read_prefetch_u8_m1k32v1(
-    long baseoffset, int width_minus_one, int height_minus_one,
+    intptr_t baseoffset, int width_minus_one, int height_minus_one,
     int pitch_minus_one, cute::intel::coord_t coord, enum CacheControl cache_control));
 SYCL_DEVICE_BUILTIN(void __builtin_IB_subgroup_block_read_prefetch_u8_m2k32v1(
-    long baseoffset, int width_minus_one, int height_minus_one,
+    intptr_t baseoffset, int width_minus_one, int height_minus_one,
     int pitch_minus_one, cute::intel::coord_t coord, enum CacheControl cache_control));
 SYCL_DEVICE_BUILTIN(void __builtin_IB_subgroup_block_read_prefetch_u8_m4k32v1(
-    long baseoffset, int width_minus_one, int height_minus_one,
+    intptr_t baseoffset, int width_minus_one, int height_minus_one,
     int pitch_minus_one, cute::intel::coord_t coord, enum CacheControl cache_control));
 SYCL_DEVICE_BUILTIN(void __builtin_IB_subgroup_block_read_prefetch_u8_m8k32v1(
-    long baseoffset, int width_minus_one, int height_minus_one,
+    intptr_t baseoffset, int width_minus_one, int height_minus_one,
     int pitch_minus_one, cute::intel::coord_t coord, enum CacheControl cache_control));
 
 #undef SYCL_DEVICE_BUILTIN
@@ -269,7 +269,7 @@ struct XE_2D_U8x1x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<ushort *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m1k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -281,7 +281,7 @@ struct XE_2D_U8x1x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u8_m1k32v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord, CacheControl::kL1C_L3C);
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord, CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
           "Trying to use block prefetch on non-PVC hardware");
@@ -302,7 +302,7 @@ struct XE_2D_U8x2x32_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort2 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m2k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -314,7 +314,7 @@ struct XE_2D_U8x2x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u8_m2k32v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord, CacheControl::kL1C_L3C);
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord, CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
           "Trying to use block prefetch on non-PVC hardware");
@@ -348,7 +348,7 @@ struct XE_2D_U8x4x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u8_m4k32v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -382,7 +382,7 @@ struct XE_2D_U8x8x32_LD_N {
                                       intel::coord_t coord) {
 #if defined(SYCL_INTEL_TARGET)
       __builtin_IB_subgroup_block_read_prefetch_u8_m8k32v1(
-          (long)baseoffset, width - 1, height - 1, pitch - 1, coord,
+          (intptr_t)baseoffset, width - 1, height - 1, pitch - 1, coord,
           CacheControl::kL1C_L3C);
 #else
       CUTE_INVALID_CONTROL_PATH(
@@ -455,7 +455,7 @@ struct XE_2D_U4x16x16_LD_T {
 #if defined(SYCL_INTEL_TARGET)
     *reinterpret_cast<intel::uint2 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transpose_u32_k2(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -474,7 +474,7 @@ struct XE_2D_U4x32x16_LD_T {
 #if defined(SYCL_INTEL_TARGET)
     *reinterpret_cast<intel::uint4 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_transpose_u32_k4(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 #else
     CUTE_INVALID_CONTROL_PATH("Trying to use block loads on non-PVC hardware");
 #endif
@@ -493,7 +493,7 @@ struct XE_2D_U4x32x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort32 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m32k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 
    // ================= shuffle begin =================
    // FIXME: the performance of shuffle algorithm here is too bad, we are working with
@@ -545,7 +545,7 @@ struct XE_2D_U4x16x64_LD_N {
     static_assert(sizeof(T) == 1, "Expected T to have size 1");
     *reinterpret_cast<intel::ushort16 *>(dst) =
         __builtin_IB_subgroup_block_read_flat_u8_m16k32v1(
-            (long)(baseoffset), width - 1, height - 1, pitch - 1, coord);
+            (intptr_t)(baseoffset), width - 1, height - 1, pitch - 1, coord);
 
    // ================= shuffle begin =================
    // FIXME: the performance of shuffle algorithm here is too bad, we are working with
